Add lit-html todo app
