+++ src/index.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Todo - Lit HTML</title>
+
+  <style>
+    body {
+      background-color: #efefef;
+      margin: 0;
+      padding: 0;
+    }
+  </style>
+</head>
+
+<body>
+  <app-main></app-main>
+
+  <script src="./scripts/main.js"></script>
+</body>
+
+</html>
